feat!: remove deprecated methods
Close #108
